Parallel_Background_Subtraction_MPI/requirements.txt
mpi4py
numpy
opencv-python
Pillow
